Conflicts:
	CHANGELOG
	cocos2d-ios.xcodeproj/project.pbxproj
	cocos2d/CCTextureCache.m

Former-commit-id: dc10b1fae69019ebfc4e7d6c0b9e246408f8bf5a

+++ cocos2d/CCTransition.m
@@ -47,7 +47,7 @@
 #import "Platforms/Mac/CCEventDispatcher.h"
 #endif
 
-const uint32_t kSceneFade = 0xFADEFADE;
+const NSInteger kSceneFade = 0xFADEFADE;
 
 
 @interface CCTransitionScene (Private)
